+++ src/lib/config.ts
@@ -58,7 +58,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "2025.8.1",
-  genVersion: "2.687.1",
-  userAgent: "speakeasy-sdk/typescript 2025.8.1 2.687.1 1.0.0 wistia",
+  sdkVersion: "2025.9.1",
+  genVersion: "2.690.1",
+  userAgent:
+    "speakeasy-sdk/typescript 2025.9.1 2.690.1 1.0.0 @wistia/wistia-api-client",
 } as const;