+++ RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v2025.8.1] .
 ### Releases
 - [NPM v2025.8.1] https://www.npmjs.com/package/wistia/v/2025.8.1 - .
+
+## 2025-09-02 14:38:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.607.0 (2.690.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2025.9.1] .
+### Releases
+- [NPM v2025.9.1] https://www.npmjs.com/package/@wistia/wistia-api-client/v/2025.9.1 - .